+++ projects/brazil/config/clean_all.R
@@ -1,0 +1,2 @@
+a <- list.files(path = "config", pattern = ".rds", full.names = T)
+file.remove(a)