+++ NEWS.md
@@ -1,10 +1,16 @@
 NEWS
 ===========
 
+
+### vein 1.0.5 (date: 2023-06-14)
+
+- ef_china_speed now has long argument to process in long format
+
+
 ### vein 1.0.4 (date: 2023-05-31)
 
 - project newer cetesb emission factors
-add projet to estimate emissions in Brazil
+add project to estimate emissions in Brazil
 by state a year
 
 ### vein 1.0.3 (date: 2023-05-05)